backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.0
google-generativeai==0.7.2
boto3==1.34.100
twilio==9.2.0
requests==2.31.0
python-dateutil==2.9.0
apscheduler==3.10.4
websockets==12.0
slowapi==0.1.9